pyproject.toml
[tool.poetry]
name = "llamaindex-sample"
version = "0.1.0"
description = ""
authors = ["lz-chen <lingzhen.chen@inmeta.no>"]
readme = "README.md"
packages = [{include = "llamaindex_sample", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
llama-index-embeddings-ollama = "^0.1.2"
llama-index-llms-ollama = "^0.1.3"
llama-index-embeddings-openai = "^0.1.9"
llama-index-postprocessor-cohere-rerank = "^0.1.6"
llama-index-llms-openai = "^0.1.19"
jupyter = "^1.0.0"
llama-index-readers-file = "^0.1.22"
docx2txt = "^0.8"
torch = "^2.3.0"
transformers = "^4.41.1"
python-pptx = "^0.6.23"
pillow = "^10.3.0"
llama-index-experimental = "^0.1.3"
phoenix = "^0.9.1"
pyvis = "^0.3.2"
tiktoken = "^0.7.0"
firecrawl-py = "^0.0.20"
prettytable = "^3.10.0"
python-dotenv = "^1.0.1"
llama-index-llms-azure-openai = "^0.1.8"
crewai = {extras = ["tools"], version = "^0.36.0"}
pydantic-settings = "^2.3.4"
langchain-core = "0.2.15"
langchain = "0.2.6"
pyautogen = "^0.2.32"
llama-agents = "^0.0.9"
llama-index-agent-openai = "^0.2.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
